+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 ### Fixed
 - lint issue due to changes in tslint 5.12.0
 - stream write function type
+- pretty streams created only when needed
 
 ### Added
 - coveralls integration
+  
+## Changed
+- refactoring of express handlers
 
 ## [1.0.3] - 2018-11-08